# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Control Plane API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /gateway/preflight:
    post:
      summary: Evaluate tool call before execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreflightRequest'
      responses:
        '200':
          description: Decision result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightResponse'
...
  /approvals/{approval_id}/decision:
    post:
      summary: Approve or reject request
      parameters:
        - in: path
          name: approval_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Decision result
...
components:
  schemas:
    PreflightRequest:
      type: object
      properties:
        session_id: { type: string }
        step_id: { type: string }
        correlation_id: { type: string }
        agent: { type: object, additionalProperties: true }
        tool_call: { type: object, additionalProperties: true }
      required: [session_id, step_id, correlation_id, agent, tool_call]
    PreflightResponse:
      type: object
      properties:
        decision: { type: string }
        decision_id: { type: string }
        matched_policy_ids:
          type: array
          items: { type: string }
        reason_code: { type: string }
        reason_text: { type: string }
        risk_tags:
          type: array
          items: { type: string }
        approval_id: { type: string }
      required: [decision, decision_id, matched_policy_ids, reason_code, reason_text, risk_tags]